Read prefix file if prefix not set.
