+++ cron/internal/controller/main.go
@@ -16,7 +16,6 @@
 package main
 
 import (
-	"bytes"
 	"context"
 	"flag"
 	"fmt"
@@ -92,42 +91,6 @@ func localFiles(filenames []string) data.Iterator {
 			panic(err)
 		}
 		i, err := data.MakeIteratorFrom(f)
-		if err != nil {
-			panic(err)
-		}
-		iters = append(iters, i)
-	}
-	iter, err := data.MakeNestedIterator(iters)
-	if err != nil {
-		panic(err)
-	}
-	return iter
-}
-
-func bucketFiles(ctx context.Context) data.Iterator {
-	var iters []data.Iterator
-
-	bucket, err := config.GetInputBucketURL()
-	if err != nil {
-		panic(err)
-	}
-	prefix, err := config.GetInputBucketPrefix()
-	if err != nil {
-		panic(err)
-	}
-
-	files, err := data.GetBlobKeysWithPrefix(ctx, bucket, prefix)
-	if err != nil {
-		panic(err)
-	}
-
-	for _, f := range files {
-		b, err := data.GetBlobContent(ctx, bucket, f)
-		if err != nil {
-			panic(err)
-		}
-		r := bytes.NewReader(b)
-		i, err := data.MakeIteratorFrom(r)
 		if err != nil {
 			panic(err)
 		}
